Move panelCount and zValue to state for rerender
If the panelRenderers prop is changed to introduce/remove panels, the component
should rerender with more or less faces. To do this, panelCount and zValue need
to be calculated dynamically instead of once at the time the component
instantiates. This means they need to be part of the state and React now
recommend a new method for calculating state from updated props using
getDerivedStateFromProps.

displayedIndex state now also uses modulus to stay within panelCount which has
the effect of reversing rotation at the end of the sequence (which I think
looks cool so I'm leaving it that way!).